+++ crates/map_scatter/CHANGELOG.md
@@ -5,8 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - 2025-10-21
+
+### Changed
+
+- Serde is now an optional dependency behind the feature `serde`.
+
 ## [0.2.0] - 2025-09-30
 
 ### Added
 
-Initial release of `map_scatter` core crate.
+- Initial release of `map_scatter` core crate.